fix: reload theme before context theme wrapper is created
